KAN-41 : Rename: Set package
