+++ src/main/java/capstone/relation/workspace/school/initializer/CustomInitEvent.java
@@ -1,4 +1,4 @@
-package capstone.relation.workspace.school;
+package capstone.relation.workspace.school.initializer;
 
 import org.springframework.context.ApplicationEvent;
