+++ src/main/java/capstone/relation/workspace/school/initializer/CustomInitListener.java
@@ -1,4 +1,4 @@
-package capstone.relation.workspace.school;
+package capstone.relation.workspace.school.initializer;
 
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
