+++ src/main/java/capstone/relation/workspace/school/initializer/DatabaseInitializer.java
@@ -1,4 +1,4 @@
-package capstone.relation.workspace.school;
+package capstone.relation.workspace.school.initializer;
 
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.context.ApplicationListener;
